requirements.txt
streamlit
google-generativeai
py2neo 
python-dotenv
